Fix bug: the timer is too slow
The timer does some operations, which takes some time but this time is
ignored while the timer runs. To fix that problem, you have to save the
time, when the timer runs, and the time, when the timer is updated, and
calculate the new time with prevValue, currentTime and startTime.